Make changes in response to feedback
